受注APIの変更に対処
