+++ backend/apps/trade_flow/views/__init__.py
@@ -1,4 +1,3 @@
-from .order_create_view import OrderCreateView
-from .order_list_view import OrderListView
+from .order_view import OrderView
 from .order_update_view import OrderUpdateView
 from .order_delete_view import OrderDeleteView
